Split out gallery context
